+++ packages/account-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sophon-labs/account-core
 
+## 2.0.7-next.4
+
+### Patch Changes
+
+- Updating hooks parameter
+
 ## 2.0.7-next.3
 
 ### Patch Changes